fix: Quest ViewModel where mix quests
- quest의 순서는 이전에 정해주고 단순히 주어진 대로 관리하도록 수정
- 가독성을 위해 extension으로 분리
